requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
Faker==13.12.0
gunicorn==20.1.0
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
